Migrate more labels endpoints
