+++ server/service/endpoint_labels.go
@@ -8,81 +8,6 @@ import (
 )
 
 ////////////////////////////////////////////////////////////////////////////////
-// Delete Label
-////////////////////////////////////////////////////////////////////////////////
-
-type deleteLabelRequest struct {
-	Name string
-}
-
-type deleteLabelResponse struct {
-	Err error `json:"error,omitempty"`
-}
-
-func (r deleteLabelResponse) error() error { return r.Err }
-
-func makeDeleteLabelEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(deleteLabelRequest)
-		err := svc.DeleteLabel(ctx, req.Name)
-		if err != nil {
-			return deleteLabelResponse{Err: err}, nil
-		}
-		return deleteLabelResponse{}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Delete Label By ID
-////////////////////////////////////////////////////////////////////////////////
-
-type deleteLabelByIDRequest struct {
-	ID uint
-}
-
-type deleteLabelByIDResponse struct {
-	Err error `json:"error,omitempty"`
-}
-
-func (r deleteLabelByIDResponse) error() error { return r.Err }
-
-func makeDeleteLabelByIDEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(deleteLabelByIDRequest)
-		err := svc.DeleteLabelByID(ctx, req.ID)
-		if err != nil {
-			return deleteLabelByIDResponse{Err: err}, nil
-		}
-		return deleteLabelByIDResponse{}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Apply Label Spec
-////////////////////////////////////////////////////////////////////////////////
-
-type applyLabelSpecsRequest struct {
-	Specs []*fleet.LabelSpec `json:"specs"`
-}
-
-type applyLabelSpecsResponse struct {
-	Err error `json:"error,omitempty"`
-}
-
-func (r applyLabelSpecsResponse) error() error { return r.Err }
-
-func makeApplyLabelSpecsEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(applyLabelSpecsRequest)
-		err := svc.ApplyLabelSpecs(ctx, req.Specs)
-		if err != nil {
-			return applyLabelSpecsResponse{Err: err}, nil
-		}
-		return applyLabelSpecsResponse{}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
 // Get Label Spec
 ////////////////////////////////////////////////////////////////////////////////
