+++ server/service/handler_test.go
@@ -138,10 +138,6 @@ func TestAPIRoutes(t *testing.T) {
 			verb: "POST",
 			uri:  "/api/v1/osquery/log",
 		},
-		{
-			verb: "DELETE",
-			uri:  "/api/v1/fleet/labels/1",
-		},
 	}
 
 	for _, route := range routes {
