--- server/service/transport_labels.go
@@ -1,36 +1,0 @@
-package service
-
-import (
-	"context"
-	"encoding/json"
-	"net/http"
-)
-
-func decodeDeleteLabelRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	name, err := stringFromRequest(r, "name")
-	if err != nil {
-		return nil, err
-	}
-	var req deleteLabelRequest
-	req.Name = name
-	return req, nil
-}
-
-func decodeDeleteLabelByIDRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	id, err := uintFromRequest(r, "id")
-	if err != nil {
-		return nil, err
-	}
-	var req deleteLabelByIDRequest
-	req.ID = uint(id)
-	return req, nil
-}
-
-func decodeApplyLabelSpecsRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	var req applyLabelSpecsRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		return nil, err
-	}
-	return req, nil
-
-}
